+++ Templates/DTL-SharePoint-FullConfig/dsc/ConfigureSQLVM.ps1
@@ -9,7 +9,7 @@ configuration ConfigureSQLVM
         [Parameter(Mandatory)] [System.Management.Automation.PSCredential]$SPSetupCreds
     )
 
-    Import-DscResource -ModuleName ComputerManagementDsc, NetworkingDsc, xActiveDirectory, SqlServerDsc, xPSDesiredStateConfiguration
+    Import-DscResource -ModuleName ComputerManagementDsc, NetworkingDsc, ActiveDirectoryDsc, SqlServerDsc, xPSDesiredStateConfiguration
 
     WaitForSqlSetup
     [String] $DomainNetbiosName = (Get-NetBIOSName -DomainFQDN $DomainFQDN)
@@ -19,8 +19,6 @@ configuration ConfigureSQLVM
     [PSCredential] $SPSetupCredsQualified = New-Object PSCredential ("${DomainNetbiosName}\$($SPSetupCreds.UserName)", $SPSetupCreds.Password)
     [PSCredential] $SQLCredsQualified = New-Object PSCredential ("${DomainNetbiosName}\$($SqlSvcCreds.UserName)", $SqlSvcCreds.Password)
     $ComputerName = Get-Content env:computername
-    [Int] $RetryCount = 30
-    [Int] $RetryIntervalSec = 30
 
     Node localhost
     {
@@ -35,76 +33,107 @@ configuration ConfigureSQLVM
         #**********************************************************
         WindowsFeature ADTools  { Name = "RSAT-AD-Tools";      Ensure = "Present"; }
         WindowsFeature ADPS     { Name = "RSAT-AD-PowerShell"; Ensure = "Present"; }
-        DnsServerAddress DnsServerAddress { Address = $DNSServer; InterfaceAlias = $InterfaceAlias; AddressFamily  = 'IPv4' }
+        
+        DnsServerAddress SetDNS { Address = $DNSServer; InterfaceAlias = $InterfaceAlias; AddressFamily  = 'IPv4' }
+
+        Firewall DatabaseEngineFirewallRule
+        {
+            Direction = "Inbound"
+            Name = "SQL-Server-Database-Engine-TCP-In"
+            DisplayName = "SQL Server Database Engine (TCP-In)"
+            Description = "Inbound rule for SQL Server to allow TCP traffic for the Database Engine."
+            Group = "SQL Server"
+            Enabled = "True"
+            Protocol = "TCP"
+            LocalPort = "1433"
+            Ensure = "Present"
+        }
 
         #**********************************************************
         # Join AD forest
         #**********************************************************
-        xWaitForADDomain DscForestWait
+        # If WaitForADDomain does not find the domain whtin "WaitTimeout" secs, it will signar a restart to DSC engine "RestartCount" times
+        WaitForADDomain DscForestWait
         {
-            DomainName           = $DomainFQDN
-            DomainUserCredential = $DomainAdminCredsQualified
-            RetryCount           = $RetryCount
-            RetryIntervalSec     = $RetryIntervalSec
-            #DependsOn            = "[DnsServerAddress]DnsServerAddress"
+            DomainName              = $DomainFQDN
+            WaitTimeout             = 1200
+            RestartCount            = 2
+            WaitForValidCredentials = $True
+            Credential              = $DomainAdminCredsQualified
+            DependsOn               = "[DnsServerAddress]SetDNS"
+        }
+
+        # WaitForADDomain sets reboot signal only if WaitForADDomain did not find domain within "WaitTimeout" secs
+        PendingReboot RebootOnWaitForADDomainSignal
+        {
+            Name             = "RebootOnWaitForADDomainSignal"
+            SkipCcmClientSDK = $true
+            DependsOn        = "[WaitForADDomain]DscForestWait"
         }
 
         Computer DomainJoin
         {
-            Name = $ComputerName
+            Name       = $ComputerName
             DomainName = $DomainFQDN
             Credential = $DomainAdminCredsQualified
-            DependsOn = "[xWaitForADDomain]DscForestWait"
+            DependsOn  = "[PendingReboot]RebootOnWaitForADDomainSignal"
+        }
+
+        PendingReboot RebootOnComputerSignal
+        {
+            Name             = "RebootOnComputerSignal"
+            SkipCcmClientSDK = $true
+            DependsOn        = "[Computer]DomainJoin"
         }
 
         #**********************************************************
         # Create accounts and configure SQL Server
         #**********************************************************
-        xADUser CreateSqlSvcAccount
+        ADUser CreateSqlSvcAccount
         {
-            DomainAdministratorCredential = $DomainAdminCredsQualified
             DomainName = $DomainFQDN
             UserName = $SqlSvcCreds.UserName
             Password = $SQLCredsQualified
             PasswordNeverExpires = $true
             Ensure = "Present"
-            DependsOn = "[Computer]DomainJoin"
+            PsDscRunAsCredential = $DomainAdminCredsQualified
+            DependsOn = "[PendingReboot]RebootOnComputerSignal"
         }
 
-        xADServicePrincipalName UpdateSqlSPN1
+        ADServicePrincipalName UpdateSqlSPN1
         {
-            Ensure               = "Present"
             ServicePrincipalName = "MSSQLSvc/$ComputerName.$($DomainFQDN):1433"
             Account              = $SqlSvcCreds.UserName
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[xADUser]CreateSqlSvcAccount"
+            Ensure               = "Present"
+            DependsOn            = "[ADUser]CreateSqlSvcAccount"
         }
 
-        xADServicePrincipalName UpdateSqlSPN2
+        ADServicePrincipalName UpdateSqlSPN2
         {
-            Ensure               = "Present"
             ServicePrincipalName = "MSSQLSvc/$ComputerName.$DomainFQDN"
             Account              = $SqlSvcCreds.UserName
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[xADUser]CreateSqlSvcAccount"
+            Ensure               = "Present"
+            DependsOn            = "[ADUser]CreateSqlSvcAccount"
         }
 
-        xADServicePrincipalName UpdateSqlSPN3
+        ADServicePrincipalName UpdateSqlSPN3
         {
-            Ensure               = "Present"
             ServicePrincipalName = "MSSQLSvc/$($ComputerName):1433"
             Account              = $SqlSvcCreds.UserName
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[xADUser]CreateSqlSvcAccount"
+            Ensure               = "Present"
+            DependsOn            = "[ADUser]CreateSqlSvcAccount"
         }
 
-        xADServicePrincipalName UpdateSqlSPN4
+        ADServicePrincipalName UpdateSqlSPN4
         {
-            Ensure               = "Present"
             ServicePrincipalName = "MSSQLSvc/$ComputerName"
             Account              = $SqlSvcCreds.UserName
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[xADUser]CreateSqlSvcAccount"
+            Ensure               = "Present"
+            DependsOn            = "[ADUser]CreateSqlSvcAccount"
         }
 
         SqlServiceAccount SetSqlInstanceServiceAccount
@@ -114,18 +143,18 @@ configuration ConfigureSQLVM
             ServiceType    = "DatabaseEngine"
             ServiceAccount = $SQLCredsQualified
             RestartService = $true
-            DependsOn      = "[xADServicePrincipalName]UpdateSqlSPN1", "[xADServicePrincipalName]UpdateSqlSPN2", "[xADServicePrincipalName]UpdateSqlSPN3", "[xADServicePrincipalName]UpdateSqlSPN4"
+            DependsOn      = "[ADServicePrincipalName]UpdateSqlSPN1", "[ADServicePrincipalName]UpdateSqlSPN2", "[ADServicePrincipalName]UpdateSqlSPN3", "[ADServicePrincipalName]UpdateSqlSPN4"
         }
 
-        xADUser CreateSPSetupAccount
+        ADUser CreateSPSetupAccount
         {
-            DomainAdministratorCredential = $DomainAdminCredsQualified
             DomainName = $DomainFQDN
             UserName = $SPSetupCreds.UserName
             Password = $SPSetupCredsQualified
             PasswordNeverExpires = $true
+            PsDscRunAsCredential = $DomainAdminCredsQualified
             Ensure = "Present"
-            DependsOn = "[Computer]DomainJoin"
+            DependsOn = "[PendingReboot]RebootOnComputerSignal"
         }
 
         SqlLogin AddDomainAdminLogin
@@ -135,7 +164,7 @@ configuration ConfigureSQLVM
             ServerName = $ComputerName
             InstanceName = "MSSQLSERVER"
             LoginType = "WindowsUser"
-            DependsOn = "[Computer]DomainJoin"
+            DependsOn = "[PendingReboot]RebootOnComputerSignal"
         }
 
         SqlLogin AddSPSetupLogin
@@ -145,7 +174,7 @@ configuration ConfigureSQLVM
             ServerName = $ComputerName
             InstanceName = "MSSQLSERVER"
             LoginType = "WindowsUser"
-            DependsOn = "[xADUser]CreateSPSetupAccount"
+            DependsOn = "[ADUser]CreateSPSetupAccount"
         }
 
         SqlRole GrantSQLRoleSysadmin
@@ -183,21 +212,7 @@ configuration ConfigureSQLVM
             ServerName   = $ComputerName
             InstanceName = "MSSQLSERVER"
             MaxDop       = 1
-            DependsOn    = "[Computer]DomainJoin"
-        }
-
-        # Open port on the firewall when everything is ready, as SharePoint DSC is testing it to start creation of the farm
-        Firewall DatabaseEngineFirewallRule
-        {
-            Direction   = "Inbound"
-            Name        = "SQL-Server-Database-Engine-TCP-In"
-            DisplayName = "SQL Server Database Engine (TCP-In)"
-            Description = "Inbound rule for SQL Server to allow TCP traffic for the Database Engine."
-            Group       = "SQL Server"
-            Enabled     = "True"
-            Protocol    = "TCP"
-            LocalPort   = "1433"
-            Ensure      = "Present"
+            DependsOn    = "[PendingReboot]RebootOnComputerSignal"
         }
     }
 }
@@ -257,7 +272,8 @@ $SPSetupCreds = Get-Credential -Credential "spsetup"
 $DNSServer = "10.0.1.4"
 $DomainFQDN = "contoso.local"
 
-ConfigureSQLVM -DNSServer $DNSServer -DomainFQDN $DomainFQDN -DomainAdminCreds $DomainAdminCreds -SqlSvcCreds $SqlSvcCreds -SPSetupCreds $SPSetupCreds -ConfigurationData @{AllNodes=@(@{ NodeName="localhost"; PSDscAllowPlainTextPassword=$true })} -OutputPath "C:\Data\\output"
-Start-DscConfiguration -Path "C:\Data\output" -Wait -Verbose -Force
+$outputPath = "C:\Packages\Plugins\Microsoft.Powershell.DSC\2.80.0.0\DSCWork\ConfigureSQLVM.0\ConfigureSQLVM"
+ConfigureSQLVM -DNSServer $DNSServer -DomainFQDN $DomainFQDN -DomainAdminCreds $DomainAdminCreds -SqlSvcCreds $SqlSvcCreds -SPSetupCreds $SPSetupCreds -ConfigurationData @{AllNodes=@(@{ NodeName="localhost"; PSDscAllowPlainTextPassword=$true })} -OutputPath $outputPath
+Start-DscConfiguration -Path $outputPath -Wait -Verbose -Force
 
 #>